requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
markdownify==0.11.6
aiohttp==3.9.1
